+++ src/egon/data/datasets/chp/__init__.py
@@ -16,7 +16,6 @@ from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 import geopandas as gpd
 import pandas as pd
-import pypsa
 
 from egon.data import config, db
 from egon.data.datasets import (
@@ -30,7 +29,6 @@ from egon.data.datasets.chp.small_chp import (
     assign_use_case,
     existing_chp_smaller_10mw,
     extension_per_federal_state,
-    extension_to_areas,
     select_target,
 )
 from egon.data.datasets.power_plants import (
@@ -39,7 +37,6 @@ from egon.data.datasets.power_plants import (
     filter_mastr_geometry,
     scale_prox2now,
 )
-from egon.data.datasets.pypsaeur import read_network
 from egon.data.metadata import (
     context,
     generate_resource_fields_from_sqla_model,
@@ -359,10 +356,11 @@ def insert_biomass_chp(scenario):
 
     """
 
-    # import target values from NEP 2021, scneario C 2035
+    # import target values from NEP 2021, scenario C 2035
+    # or from NEP2025, scenario C 2037/2045
     target = select_target("biomass", scenario)
 
-    # import data for MaStR
+    # import data from MaStR
     mastr = pd.read_csv(Chp.sources.files["mastr_biomass"]).query(
         "EinheitBetriebsstatus=='InBetrieb'"
     )
@@ -380,8 +378,9 @@ def insert_biomass_chp(scenario):
             ).states.values
         )
     ]
-    # Scaling will be done per federal state in case of eGon2035 scenario.
-    if scenario == "eGon2035":
+    # Scaling is done per federal state in case of eGon2035, reGon2037 and 
+    # reGon2045 scenario.
+    if scenario in ["eGon2035", "reGon2037", "reGon2045"]:
         level = "federal_state"
     else:
         level = "country"
@@ -399,6 +398,8 @@ def insert_biomass_chp(scenario):
         mastr_loc = assign_bus_id(mastr_loc, Chp.sources)
     mastr_loc = assign_use_case(mastr_loc, Chp.sources, scenario)
 
+    nep_version = "NEP 2021" if scenario == "eGon2035" else "NEP 2025"
+
     # Insert entries with location
     session = sessionmaker(bind=db.engine())()
     for i, row in mastr_loc.iterrows():
@@ -406,7 +407,7 @@ def insert_biomass_chp(scenario):
             entry = EgonChp(
                 sources={
                     "chp": "MaStR",
-                    "el_capacity": "MaStR scaled with NEP 2021",
+                    "el_capacity": f"MaStR scaled with {nep_version}",
                     "th_capacity": "MaStR",
                 },
                 source_id={"MastrNummer": row.EinheitMastrNummer},
@@ -423,9 +424,9 @@ def insert_biomass_chp(scenario):
     session.commit()
 
 
-def insert_chp_statusquo(scn="status2019"):
+def insert_chp_statusquo(scn="status2024"):
 
-    # import data for MaStR
+    # import data from MaStR
     mastr = pd.read_csv(Chp.sources.files["mastr_combustion"])
 
     mastr_biomass = pd.read_csv(Chp.sources.files["mastr_biomass"])
@@ -473,8 +474,8 @@ def insert_chp_statusquo(scn="status2019"):
         GROUP BY gen"""
     ).set_index("gen")
 
-    # Assing Laengengrad and Breitengrad to chps without location data
-    # based on the centroid of the municipaltiy
+    # Assigning degree of longitude (Laengengrad) and latitude (Breitengrad) to 
+    # chps without location data based on the centroid of the municipaltiy
     idx_no_location = mastr[
         (mastr.Laengengrad.isnull())
         & (mastr.Gemeinde.isin(geom_municipalities.index))
@@ -551,43 +552,47 @@ def insert_chp_statusquo(scn="status2019"):
     session.commit()
 
 
-def insert_chp_egon2035():
-    """Insert CHP plants for eGon2035 considering NEP and MaStR data
+def insert_chp():
+    """Insert CHP plants for eGon2035, reGon2037 and reGon2045 considering NEP and MaStR data
 
     Returns
     -------
     None.
 
     """
-
-    insert_biomass_chp("eGon2035")
-
-    # Insert large CHPs based on NEP's list of conventional power plants
-    MaStR_konv = insert_large_chp(
-        Chp.sources, Chp.targets.tables["chp_table"], EgonChp
-    )
-
-    # Insert smaller CHPs (< 10MW) based on existing locations from MaStR
-    existing_chp_smaller_10mw(Chp.sources, MaStR_konv, EgonChp)
-
-    gpd.GeoDataFrame(
-        MaStR_konv[
-            [
-                "EinheitMastrNummer",
-                "el_capacity",
-                "geometry",
-                "carrier",
-                "plz",
-                "city",
-                "federal_state",
+    
+    # filters scenarios for all non-status-scenarios
+    scenarios = [s for s in config.settings()["egon-data"]["--scenarios"] if "status" not in str(s).lower()]
+    for scenario in scenarios:
+        insert_biomass_chp(scenario)
+    
+        # Insert large CHPs based on NEP's list of conventional power plants
+        MaStR_konv = insert_large_chp(
+            Chp.sources, Chp.targets.tables["chp_table"], EgonChp,
+            scenario
+        )
+    
+        # Insert smaller CHPs (< 10MW) based on existing locations from MaStR
+        existing_chp_smaller_10mw(Chp.sources, MaStR_konv, EgonChp, scenario)
+    
+        gpd.GeoDataFrame(
+            MaStR_konv[
+                [
+                    "EinheitMastrNummer",
+                    "el_capacity",
+                    "geometry",
+                    "carrier",
+                    "plz",
+                    "city",
+                    "federal_state",
+                ]
             ]
-        ]
-    ).to_postgis(
-        Chp.targets.get_table_name("mastr_conventional_without_chp"),
-        schema=Chp.targets.get_table_schema("mastr_conventional_without_chp"),
-        con=db.engine(),
-        if_exists="replace",
-    )
+        ).to_postgis(
+            Chp.targets.get_table_name("mastr_conventional_without_chp"),
+            schema=Chp.targets.get_table_schema("mastr_conventional_without_chp"),
+            con=db.engine(),
+            if_exists="replace",
+        )
 
 
 def extension_BW():
@@ -654,105 +659,27 @@ def extension_SH():
     extension_per_federal_state("SchleswigHolstein", EgonChp)
 
 
-def insert_chp_egon100re():
-    """Insert CHP plants for eGon100RE considering results from pypsa-eur-sec
-
-    Returns
-    -------
-    None.
-
-    """
-
-    db.execute_sql(f"""
-        DELETE FROM {Chp.targets.tables['chp_table']}
-        WHERE scenario = 'eGon100RE'
-        """)
-
-    # select target values from pypsa-eur-sec
-    additional_capacity = db.select_dataframe(f""" # noqa: S608
-        SELECT capacity
-        FROM {Chp.sources.tables['scenario_capacities']}
-        WHERE scenario_name = 'eGon100RE'
-        AND carrier = 'urban_central_gas_CHP'
-        """).capacity[0]
-
-    if config.settings()["egon-data"]["--dataset-boundary"] != "Everything":
-        additional_capacity /= 16
-
-    network = read_network()
-
-    chp_index = "DE0 0 urban central gas CHP-2045"
-
-    standard_chp_th = 10
-    standard_chp_el = (
-        standard_chp_th
-        * network.links.loc[chp_index, "efficiency"]
-        / network.links.loc[chp_index, "efficiency2"]
-    )
-
-    areas = db.select_geodataframe(f"""
-            SELECT
-            residential_and_service_demand as demand, area_id,
-            ST_Transform(ST_PointOnSurface(geom_polygon), 4326)  as geom
-            FROM
-            {Chp.sources.tables['district_heating_areas']}
-            WHERE scenario = 'eGon100RE'
-            """)
-
-    existing_chp = pd.DataFrame(
-        data={
-            "el_capacity": standard_chp_el,
-            "th_capacity": standard_chp_th,
-            "voltage_level": 5,
-        },
-        index=range(1),
-    )
-
-    flh = (
-        network.links_t.p0[chp_index].sum()
-        / network.links.p_nom_opt[chp_index]
-    )
-
-    extension_to_areas(
-        areas,
-        additional_capacity,
-        existing_chp,
-        flh,
-        EgonChp,
-        district_heating=True,
-        scenario="eGon100RE",
-    )
-
-
 tasks = (create_tables,)
 
 insert_per_scenario = set()
 
-if "status2019" in config.settings()["egon-data"]["--scenarios"]:
+scenarios = config.settings()["egon-data"]["--scenarios"]
+  
+if "status2024" in scenarios:
     insert_per_scenario.add(
         wrapped_partial(
-            insert_chp_statusquo, scn="status2019", postfix="_2019"
+            insert_chp_statusquo, scn="status2024", postfix="_2024"
         )
     )
 
-if "status2023" in config.settings()["egon-data"]["--scenarios"]:
-    insert_per_scenario.add(
-        wrapped_partial(
-            insert_chp_statusquo, scn="status2023", postfix="_2023"
-        )
-    )
-
-if "eGon2035" in config.settings()["egon-data"]["--scenarios"]:
-    insert_per_scenario.add(insert_chp_egon2035)
-
-if "eGon100RE" in config.settings()["egon-data"]["--scenarios"]:
-    insert_per_scenario.add(insert_chp_egon100re)
+if any(s in scenarios for s in ["eGon2035", "reGon2037", "reGon2045"]):
+    insert_per_scenario.add(insert_chp)
 
 tasks = tasks + (insert_per_scenario, assign_heat_bus)
 
 extension = set()
 
-if "eGon2035" in config.settings()["egon-data"]["--scenarios"]:
+if any(s in scenarios for s in ["eGon2035", "reGon2037", "reGon2045"]):
     # Add one task per federal state for small CHP extension
     if (
         config.settings()["egon-data"]["--dataset-boundary"]
@@ -789,7 +716,7 @@ class Chp(Dataset):
 
     sources = DatasetSources(
         tables={
-            "list_conv_pp": "supply.egon_nep_2021_conventional_powerplants",
+            "list_conv_pp": "supply.egon_nep_conventional_powerplants",
             "egon_mv_grid_district": "grid.egon_mv_grid_district",
             "ehv_voronoi": "grid.egon_ehv_substation_voronoi",
             "etrago_buses": "grid.egon_etrago_bus",
@@ -817,7 +744,7 @@ class Chp(Dataset):
     Extract combined heat and power plants for each scenario
 
     This dataset creates combined heat and power (CHP) plants for each scenario and defines their use case.
-    The method bases on existing CHP plants from Marktstammdatenregister. For the eGon2035 scenario,
+    The method is based on existing CHP plants from Marktstammdatenregister. For the eGon2035 scenario,
     a list of CHP plans from the grid operator is used for new largescale CHP plants. CHP < 10MW are
     randomly distributed.
     Depending on the distance to a district heating grid, it is decided if the CHP is used to
@@ -825,7 +752,6 @@ class Chp(Dataset):
 
 
     *Dependencies*
-      * :py:class:`GasAreaseGon100RE <egon.data.datasets.gas_areas.GasAreaseGon100RE>`
       * :py:class:`GasAreaseGon2035 <egon.data.datasets.gas_areas.GasAreaseGon2035>`
       * :py:class:`DistrictHeatingAreas <egon.data.datasets.district_heating_areas.DistrictHeatingAreas>`
       * :py:class:`IndustrialDemandCurves <egon.data.datasets.industry.IndustrialDemandCurves>`
@@ -844,7 +770,7 @@ class Chp(Dataset):
     #:
     name: str = "Chp"
     #:
-    version: str = "0.0.12"
+    version: str = "0.0.13"
 
     def __init__(self, dependencies):
         super().__init__(